requirements.txt
deap==1.3.3
google_api_python_client==2.86.0
joblib==1.2.0
keras==2.9.0
mediapipe==0.9.3.0
mediapipe_silicon==0.8.11
numpy==1.23.5
opencv_contrib_python==4.7.0.68
protobuf==4.22.3
PuLP==2.7.0
python-dotenv==1.0.0
pytube==12.1.3
scikit_learn==1.2.2
spotipy==2.23.0
tensorflow==2.12.0
tensorflow_macos==2.9.0